Refactor comments into seperate controller
